requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.10.1
python-telegram-bot==21.10
requests==2.32.3
SQLAlchemy==2.0.37
